shell: make it convenient to have gardens and trees with the same name
Automatically chdir into a specific tree when the tree name exactly matches
the tree expression that was used to find it.

This is useful when the a garden exists with the same name as a
tree, and the garden also contains that same tree.  Using the
common name as a tree expression finds the matching tree inside
the garden results and selects it as the working directory for
the shell.