cmd: allow passing additional arguments to custom commands
Forward command-line arguments that are specified after the "--"
double-dash end-of-options marker to custom commands.

Commands have the arguments available via the traditional "$@" shell
syntax.  This is supported by the shell itself so the eval logic must be
careful to leave $0 and other digit variables unresolved so that the
shell can see them.