cmd: resolve command paths in the resolved environment
Some commands need to be absolute-path resolved before handing off to
subprocess because they won't be found by popen.  Popen does allow
passing PATH environment, which allows commands to be found from
children of the sub-command, but the environment is not consulted when
resolving the command.

Teach exec_in_context() to find commands in the resolved environment.