garden grow: add support for configuring the branch to clone
Implement the "branches" setting for trees.
Teach the "grow" command to clone specific branches when the "branch"
setting is configured for a tree.

This is another step towards git-worktree support. This is syntax
that is common between worktrees and regular trees and sets us
up towards implementing support for git-worktree.

Related-to: #1
Signed-off-by: David Aguilar <davvid@gmail.com>