config: refactor Template and Tree interactions
Change the guts of Template so that it embed a Tree struct.
This allows us to avoid having any duplicate fields between Template
and Tree structs.

Move all of the updating and management of tree-related variables
into the Tree and Template structs.

Add fields related to Bare Repositories and Git Worktrees to the
Tree struct and read them in the config reader.

The Git Worktree fields are now available so that commands such as
"garden grow" can start to use them.

Related-to: #1
Signed-off-by: David Aguilar <davvid@gmail.com>