doc: update the worktree examples
Experimentation with "git worktree" shows that the only sensible way
to use it is to start from a shared regular clone, not a bare repository.
This is for various reasons.

One simple reason is that doing "git fetch" into a "git clone --bare"-cloned
repository updates branches in place, which can lead to surprising behavior
in worktrees for users that are not aware of what is happening.

Another complication is that "git config core.bare" returns true in the
git-worktree created worktrees, as they share the same .gitconfig.

It is clear that "git worktree" was not designed to be used alongside
bare repositories. Change our documentation to follow suite so that
we can finish implementing the worktree support following this
paradigm as the main use case.

Related-to: #1