cmd: teach "garden cmd" to run commands in breadth-first order
Make it possible to run all commands on a tree before continuing on
to the next tree. This is a complement to the default depth-first traversal
which runs each command across all trees before running the next command.

Add a -b/--breadth-first option to "garden cmd" for this traversal mode.
Document the depth-first and breadth-first traversal modes.

Closes #3
Signed-off-by: David Aguilar <davvid@gmail.com>