query: add shared_worktree_path() to identify shared worktrees
We can skip some operations if we have a way to detect when we have already
configured a shared git worktree. Add shared_worktree_path() to calculate
a path that can be used as a hash key that folds parent and child
repositories created by "git worktree" into a single value.

Signed-off-by: David Aguilar <davvid@gmail.com>