cmd: run commands under "zsh -e -c ..." by default
Make it so that commands with multiple statements halt execution
when the first non-zero exit status is encountered. We do this by
passing the "-e" option to the configured shell.

Add a --no-errexit option to opt-out of this behavior.
We can disable this on a per-command basis by adding "set +e"
to the top of the command, so there is little downside to
making this change.

Signed-off-by: David Aguilar <davvid@gmail.com>