ideas: remove the entries related to modular/reusable configuration
The subsequent commits will add support for `garden.includes` so these
todo entries are no longer needed.

This brings us much closer to something we could call v1.0.0. The
"grafts" feature, in hindsight, may not really be needed if we wanted to
keep things simple, so there's a 50/50 chance that it won't be added.
TBD.

For now, having support for "includes" seems both powerful and simple to
implement. We'll see whether we ever end up needing grafts.

Signed-off-by: David Aguilar <davvid@gmail.com>