workflow: replace "make" with "garden"
Provide a "garden.yaml" file to handle development workflow commands.
Garden is now self-hosted and no longer uses a "Makefile".

The installation instructions delegate to "cargo install" so there
is no need to provide a "make install" target anymore.

"cargo build" and "cargo test" is how we can build and test the code
when "garden" is not installed, so there is little reason to keep
the "Makefile" around anymore.

Update the CI and installation instructions to use "garden" and "cargo"
instead of "make".

Signed-off-by: David Aguilar <davvid@gmail.com>