doc: document how to fix the zsh completion for "garden cmd"
`garden cmd <TAB>` generates an error on zsh:

    _arguments:comparguments:325: doubled rest argument definition:
    *::arguments -- Arguments to forward to custom commands:

The root cause of the issue is that the generated completion script
contains two `*::` entries in the `garden cmd` arguments:

    '*::commands -- Custom commands to run over the resolved trees:' \
    '*::arguments -- Arguments to forward to custom commands:' \

A simple workaround is to drop the 2nd `*::arguments` entry.
Use a `grep` filter to workaround the zsh completion bug from
clap-rs/clap#3022 for now.

Document the workaround and add link to the issue.

Closes #10
Related-to: https://github.com/clap-rs/clap/issues/3022
Signed-off-by: David Aguilar <davvid@gmail.com>