eval: use shellexpand::full_with_context_no_errors()
It is better to evaluate unknown variables to empty strings across the
board for consistency. Graft variable expressions are currently a
work in progress and unsupported, but their code path was the only
one that returned an error.

Returning an error is less desirable because it prevents the entire
expression from being evaluated rather than just the single missing
variable. It can be argued that garden should support a "strict"
mode that does not allow undefined variables, but they are useful
when building shell snippets so it would be a tradeoff. The current
behavior is non-strict.

Simplify the code by returning empty strings in these scenarios instead.
This allows us to use the simpler shellexpand::full_with_context_no_errors()
function instead.

Undefined ${graft::variables} alongside defined ${variables} will now
evaluate instead of erroring-out and not evaluate anything.

We can always go back to using shellexpand::full_with_context() in the
future if we ever end up needing a "strict" mode that does not allow
${undefined} garden variable references.

Signed-off-by: David Aguilar <davvid@gmail.com>