commands: make commands follow a "last one wins" rule
Allow commnads to override themselves when they are defined in
multiple locations. Previously garden would aggregate command
definitions.

This had the effect of making it so that commands
defined via include files would be retained and run in addition
to commands defined in the top-level "garden.yaml". Furthermore,
if multiple include files defined the same command then all of
the commands would be run rather than just the last one.

The new behavior is that commands that are seen later
override commands that were seen earlier in the include list.
This is effectively a "last one wins" rule ~ when two includes define
the same command then only the latter will be used.

If both "garden.includes" and the top-level "garden.yaml"
define the same command then only the one defined in the
top-level "garden.yaml" will be used. The top-level "garden.yaml"
has the highest priority when defining commands.

Closes #14
Signed-off-by: David Aguilar <davvid@gmail.com>