model: allow trees to override template-provided commands
Store commands in a HashMap so that they are naturally deduplicated.
Make further progress towards uniformly handling overrides of commands
and other fields by changing the data structure used for commands.

Use a HashMap so that a command can only have a single definition.
This makes it so that commands defined in both a tempalte and a tree
will be overridden by the tree-provided definition.

Related-to: #14
Signed-off-by: David Aguilar <davvid@gmail.com>