model: use a VariableHashMap to store variables
Simplify how the top-level variables are stored to follow a simple last-one-wins
override behavior. Variables now behave more or less the same way commands do
as far as being able to override and replace them in a garden file.

This only updates how the top-level variables are stored. The next step
is to update how the tree-level variables to follow the same design.

Related-to: #14
Signed-off-by: David Aguilar <davvid@gmail.com>