cmd: set PWD to the specified directory in exec_in_dir()
Guarantee that PWD variable and the current directory are always in sync.

I audited the subprocess crate and verified that it does not
set PWD on our behalf. Most of the time we won't notice if $PWD is
not set to the current directory because we run commands through a
shell, and the shell updates $PWD on our behalf.

Make the behavior consistent in all scenarios by updating PWD
in the subprocess environment created by exec_in_dir().
This avoids issues in commands that are sensitive to $PWD.

The following links are related to the subprocess module in Python
where users noticed the discrepancy.

Related-to: https://bugs.python.org/issue4057
Related-to: https://bugs.python.org/issue30798