config: create an implicit "." tree when "trees" is empty
Make it easier to use "garden" as a simple command runner by making the
"trees" block optional.

If the garden file provides a "commands" block then it is now possible
to omit the "trees" block entirely. A default tree called "." will be
synthesized into existence with its path set to "${GARDEN_CONFIG_DIR}".