eval: improve support for evaluating TREE_PATH in environment blocks
Add custom "environment" evaluator methods that are able to evaluate just a
single variable. This is needed to prevent the evaluation machinery from
running into infinite loops due to the fact that environment() evaluates
all variables, which effectively introduces circular dependencies any
time any variable is used.

The environment machinery still needs further work. A follow-up
approach can resolve garden environments in two passes. The first
pass can gather all of the relevant trees and then a 2nd pass
can perform evaluation using those trees. The current approach
does not fully handle grafted trees because the tree context is
wired up to either contain the root config, which allows the
garden to be found but does not find the grafted tree, or it
can contain the grafted configuration which allows the grafted
tree to be found but fails to find the garden.