garden: add support for custom shell commands
garden.shell can not be configured to arbitrary interpreter commands.
We now run the value through the "shlex" parser to separate the
string into individual command-line arguments.

Shorthand shell values are still supported and special-cased.
Only when the command string expands to a command with multiple
arguments we treat the command like a custom shell and disable the
internal builtin arguments.