grow: properly clone into empty directories
Using "git submodule" can result in empty directories on disk
that we may want to associate with a garden tree.

We can safely clone into empty directories but we need to
not consider empty directories as being grown trees.

Check for empty directories and the existence of ".git"
to flag trees as valid.