model: use UnsafeCell as the Arena storage
The RefCell that was being used to store it can be made to panic when
running "garden exec" in parallel mode. We only need to be able to clone
the arena by reading from it in a read-only fashion, so use of
UnsafeCell lets us avoid RefCell's limitations and properly implement
the thread-safe clone() for ApplicationContext.